Add BubbleTail animations
